# storage: use only one mutex for store struct