@gyuho 2015-12-21
For future reference, test script:

```sh

current_dir=$(pwd)

dot="."
test_function_name="XXX"
# benchmark_function_name="BenchmarkStorePut"
benchmark_function_name="BenchmarkStoreTxnPut"

repeat_size=3

echo "Copying original file..."
cp $HOME/kvstore_old.go $current_dir/kvstore.go

for i in `seq 1 $repeat_size`;
do
	echo "echo 1 > /proc/sys/vm/drop_caches" | sudo sh && cd $HOME/go/pkg && rm -rf *;
	cd $current_dir && fpath="$(echo $current_dir)/old_$i.txt";
	if [ "$benchmark_function_name" == "$dot" ]; then
		echo "running all benchmarks...";
		go test -run $test_function_name -bench . -benchmem -cpu 1,4> $fpath;
	else
		echo "running only $benchmark_function_name";
		go test -run $test_function_name -bench $benchmark_function_name -benchmem -cpu 1,4 > $fpath;
	fi
done

echo "Copying new file..."
cp $HOME/kvstore_new.go $current_dir/kvstore.go

for i in `seq 1 $repeat_size`;
do
	echo "echo 1 > /proc/sys/vm/drop_caches" | sudo sh && cd $HOME/go/pkg && rm -rf *;
	cd $current_dir && fpath="$(echo $current_dir)/new_$i.txt";
	if [ "$benchmark_function_name" == "$dot" ]; then
		echo "running all benchmarks...";
		go test -run $test_function_name -bench . -benchmem -cpu 1,4 > $fpath;
	else
		echo "running only $benchmark_function_name";
		go test -run $test_function_name -bench $benchmark_function_name -benchmem -cpu 1,4 > $fpath;
	fi
done

echo "$(echo $benchmark_function_name):" > $current_dir/benchmark_results.txt;
for i in `seq 1 $repeat_size`;
do
	echo "[$(echo $i)]:" >> $current_dir/benchmark_results.txt;
	old_path="$(echo $current_dir)/old_$i.txt" && new_path="$(echo $current_dir)/new_$i.txt" && \
	benchcmp $old_path $new_path >> $current_dir/benchmark_results.txt && \
	echo "" >> $current_dir/benchmark_results.txt && \
	echo "" >> $current_dir/benchmark_results.txt;
done

echo "Done"
```