@gyuho 2015-12-21
For the `BenchmakrStorePut`, I think test results vary. Another run:

```
BenchmarkStorePut:
[1]:
benchmark               old ns/op     new ns/op     delta
BenchmarkStorePut       52914         53201         +0.54%
BenchmarkStorePut-4     60608         71636         +18.20%

benchmark               old allocs     new allocs     delta
BenchmarkStorePut       26             18             -30.77%
BenchmarkStorePut-4     26             22             -15.38%

benchmark               old bytes     new bytes     delta
BenchmarkStorePut       2000          1597          -20.15%
BenchmarkStorePut-4     2024          1787          -11.71%


[2]:
benchmark               old ns/op     new ns/op     delta
BenchmarkStorePut       43324         51032         +17.79%
BenchmarkStorePut-4     101002        62680         -37.94%

benchmark               old allocs     new allocs     delta
BenchmarkStorePut       21             26             +23.81%
BenchmarkStorePut-4     23             21             -8.70%

benchmark               old bytes     new bytes     delta
BenchmarkStorePut       1725          2000          +15.94%
BenchmarkStorePut-4     1863          1762          -5.42%


[3]:
benchmark               old ns/op     new ns/op     delta
BenchmarkStorePut       54043         35646         -34.04%
BenchmarkStorePut-4     77141         58240         -24.50%

benchmark               old allocs     new allocs     delta
BenchmarkStorePut       26             21             -19.23%
BenchmarkStorePut-4     26             22             -15.38%

benchmark               old bytes     new bytes     delta
BenchmarkStorePut       1999          1724          -13.76%
BenchmarkStorePut-4     2018          1791          -11.25%

```

All tests were run with:

```
model name	: Intel(R) Core(TM) i7-5600U CPU @ 2.60GHz
go version go1.5.2 linux/amd64
linux kernel: 3.19.0-42-generic
```