@xiang90 2015-12-21
@gyuho We have two mutex since we want to allow concurrent range query. The current implementation is bad and does not allow concurrent range query anyway. I am fine with removing the additional mutex now. But we also need to keep in mind that we might need to bring it back some day.