@gyuho 2015-12-21
> But we also need to keep in mind that we might need to bring it back some day.

For sure, I will keep that in mind. Just pushed changes addressing your suggestions. PTAL.

Thanks,